Fix bug: Cactus speeds up on each jump
This happened because the start button remains
focussed after clicking start. So each spacebar
hit to trigger jump() also triggered gameCourse(),
which has a setInterval in it. Every call to
setInterval creates new intervals, which decrement
the cactus gridColumn value by their own timers.

To fix this, the start button is disabled after
clicking it. The game has started anyway.